Update README.md
NASA Nearest Earth Objects (1910-2024) (Asteroid (4179) Tautatis is a potentially dangerous object passing at a distance of 2.3 times the distance from the Moon. (available in our dataset)) There are many dangerous objects in space, one of them is N.E.O.L. - "Nearest Earth Objects".
Some such bodies really pose a danger to the planet Earth, NASA classifies them as "is_hazardous".
This dataset contains ALL NASA observations of similar objects from 1910 to 2024!!!
There are 338,199 records of N.E.O. in the Dataset!
We tried to predict "is_hazardous" as accurately as possible! (otherwise we will not be ready for an asteroid attack).